LayoutTests/imported/w3c/web-platform-tests/svg/text/scripted/textpath-href-path-mutated-expected.txt

Text

FAIL Mutation of 'd' on a 'display: none' <path> referenced by <textPath> assert_equals: after mutation expected 50 but got 25
